Extract FormattingOptions
