+++ feature/metis/conversation/src/main/kotlin/de/tum/informatics/www1/artemis/native_app/feature/metis/conversation/ui/reply/MarkdownStyle.kt
@@ -1,6 +1,6 @@
 package de.tum.informatics.www1.artemis.native_app.feature.metis.conversation.ui.reply
 
-enum class MarkdownStyle(val startTag: String, val endTag: String) {
+internal enum class MarkdownStyle(val startTag: String, val endTag: String) {
     Bold("**", "**"),
     Italic("*", "*"),
     Underline("<ins>", "</ins>"),
